requirements.txt
astroid==1.5.3
backports.functools-lru-cache==1.4
configparser==3.5.0
enum34==1.1.6
isort==4.2.15
kafka-python==1.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
meld3==1.0.2
pylint==1.7.2
singledispatch==3.4.0.3
six==1.10.0
supervisor==3.3.3
wrapt==1.10.11
